feat: replace plan store with subscription store
