+++ frontend/src/components/DatabaseOverviewPanel.vue
@@ -211,7 +211,7 @@ export default {
     );
 
     const hasDataSourceFeature = computed(() =>
-      store.getters["plan/feature"]("bb.feature.data-source")
+      store.getters["subscription/feature"]("bb.feature.data-source")
     );
 
     const tableList = computed(() => {
