+++ frontend/src/components/ProjectMemberPanel.vue
@@ -117,7 +117,7 @@ export default {
     });
 
     const hasAdminFeature = computed(() =>
-      store.getters["plan/feature"]("bb.feature.rbac")
+      store.getters["subscription/feature"]("bb.feature.rbac")
     );
 
     const allowAddMember = computed(() => {
