+++ frontend/src/components/ProjectMemberTable.vue
@@ -112,7 +112,7 @@ export default {
     const currentUser = computed(() => store.getters["auth/currentUser"]());
 
     const hasAdminFeature = computed(() =>
-      store.getters["plan/feature"]("bb.feature.rbac")
+      store.getters["subscription/feature"]("bb.feature.rbac")
     );
 
     const state = reactive<LocalState>({});
