+++ frontend/src/components/ProjectSelect.vue
@@ -114,7 +114,7 @@ export default defineComponent({
     watchEffect(prepareProjectList);
 
     const hasAdminFeature = computed(() =>
-      store.getters["plan/feature"]("bb.feature.rbac")
+      store.getters["subscription/feature"]("bb.feature.rbac")
     );
 
     const projectList = computed((): Project[] => {
