+++ frontend/src/views/ProfileDashboard.vue
@@ -192,7 +192,7 @@ export default {
     );
 
     const hasAdminFeature = computed(() =>
-      store.getters["plan/feature"]("bb.feature.rbac")
+      store.getters["subscription/feature"]("bb.feature.rbac")
     );
 
     const principal = computed(() => {
